--- libs/mqtt/ConnectReturnCode.ts
@@ -1,16 +1,0 @@
-namespace mqtt {
-/**
- * Connect Return code
- * http://docs.oasis-open.org/mqtt/mqtt/v3.1.1/os/mqtt-v3.1.1-os.html#_Toc385349256
- */
-export const enum ConnectReturnCode {
-    Unknown = -1,
-    Accepted = 0,
-    UnacceptableProtocolVersion = 1,
-    IdentifierRejected = 2,
-    ServerUnavailable = 3,
-    BadUserNameOrPassword = 4,
-    NotAuthorized = 5
-}
-
-}
