--- libs/mqtt/ControlPacketType.ts
@@ -1,23 +1,0 @@
-namespace mqtt {
-/**
- * MQTT Control Packet type
- * http://docs.oasis-open.org/mqtt/mqtt/v3.1.1/os/mqtt-v3.1.1-os.html#_Toc353481061
- */
-export const enum ControlPacketType {
-    Connect = 1,
-    ConnAck = 2,
-    Publish = 3,
-    PubAck = 4,
-    // PubRec = 5,
-    // PubRel = 6,
-    // PubComp = 7,
-    Subscribe = 8,
-    SubAck = 9,
-    Unsubscribe = 10,
-    UnsubAck = 11,
-    PingReq = 12,
-    PingResp = 13,
-    Disconnect = 14
-}
-
-}
